LPS-119169 Extract base class
to fix initialization problems introduced in a3e060d7c9fca19259817a6745cf5a5a154ba7ac
